add argparse for easy configuration
